requirements.txt
pygame-ce
pytweening